move eth vault side logic to separate file
